move to partition package
