+++ hazelcast/src/main/java/com/hazelcast/internal/partition/ChunkSupplier.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.map.impl;
+package com.hazelcast.internal.partition;
 
 import com.hazelcast.spi.impl.operationservice.Operation;
